src/requirements.txt
asgiref==3.5.0
attrs==21.4.0
black==22.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.0.3
cryptography==36.0.1
defusedxml==0.7.1
Django==3.2.11
django-allauth==0.47.0
django-crispy-forms==1.14.0
django-debug-toolbar==3.2.4
flake8==4.0.1
idna==3.3
iniconfig==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
oauthlib==3.2.0
packaging==21.3
pathspec==0.9.0
platformdirs==2.4.1
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.3.0
pyparsing==3.0.7
pytest==6.2.5
pytest-django==4.5.2
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.0
typing_extensions==4.0.1
urllib3==1.26.8
whitenoise==5.3.0
